backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
dataclasses==0.6
Django==5.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.7.0
ecdsa==0.19.1
idna==3.10
joblib==1.4.2
mongoengine>=0.30.0rc1
numpy==2.2.4
pandas==2.2.3
pyasn1==0.6.1
PyJWT==2.10.1
pymongo>=4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2025.2
requests==2.31.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.16.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.4.0
